code/ai-compute-sever/requirements.txt
dlib >= 19.9.0
numpy >= 1.13.1
scipy >= 0.18.0
pymongo == 4.0.2
torch==1.11.0
torchaudio==0.11.0
dnspython == 2.2.1
opencv-python == 4.5.5.64
boto3 == 1.21.13
Pillow >= 9.0
requests==2.27.1
TTS==0.6.1